+++ shims.js
@@ -1,10 +1,17 @@
-if (! (URL.canParse instanceof Function)) {
-	URL.canParse = function(url, base) {
+if (!(URL.canParse instanceof Function)) {
+	URL.canParse = function (url, base) {
+		return URL.parse(url, base) instanceof URL;
+	};
+}
+
+if (!(URL.parse instanceof Function)) {
+	URL.parse = function parse(url, base) {
 		try {
-			new URL(url, base);
-			return true;
+			return new URL(url, base);
 		} catch {
-			return false;
+			{
+				return null;
+			}
 		}
 	};
 }